Refactored. Code compiles
