+++ src/contrib/mlir/core/compiler.hpp
@@ -75,10 +75,6 @@ namespace ngraph
                 {
                     return m_module;
                 }
-                // REMOVE
-                /// Executes a pre-compiled subgraph
-                //void run(std::vector<void*>& externalTensors);
-
             private:
                 struct TensorInfo
                 {
@@ -89,16 +85,6 @@ namespace ngraph
             private:
                 void buildNgDialectModule();
                 void optimizeNgDialect();
-                // REMOVE
-//                void lowerNgDialect();
-                // REMOVE
-                //void optimize();
-                // REMOVE
-                //void bindArguments(std::vector<void*>& externalTensors);
-                // REMOVE
-                //void execute();
-                // REMOVE
-                //void cleanup();
 
                 mlir::Type getMlirType(const descriptor::Tensor* tensor);
                 mlir::Type getMlirType(const element::Type& type);
@@ -126,10 +112,6 @@ namespace ngraph
 
                 void createReturn();
 
-                // REMOVE to utils
-                /// Helper to dump MLIR module into llvm::dbgs prepended by the message \p msg.
-                void dumpMlirModule(const std::string msg);
-
                 /// Converts nGraph shape-like types \p ng_shape to MLIR shape \p mlir_shape.
                 template <typename T>
                 void getMlirShape(T ngShape, llvm::SmallVectorImpl<int64_t>& mlirShape);
@@ -142,22 +124,12 @@ namespace ngraph
                 // Sub-graph to be compiled and executed with MLIR.
                 const ngraph::op::CompiledKernel* m_compiledKernel;
 
-                // REMOVE
-                // Pointers to externally allocated memory for sub-graph's input and output tensors.
-                //std::vector<void*>* m_externalTensors;
-
-                // REMOVE
-                // Arguments for the MLIR function generated for the nGraph sub-graph.
-                //llvm::SmallVector<void*, 8> m_invokeArgs;
-
                 // MLIR context that holds all the MLIR information related to the sub-graph
                 // compilation.
                 mlir::MLIRContext m_context;
 
                 mlir::OwningModuleRef m_module;
                 std::unique_ptr<mlir::OpBuilder> m_builder;
-                // REMOVE
-                //std::unique_ptr<mlir::ExecutionEngine> m_engine;
 
                 using TensorToInfo = std::pair<descriptor::Tensor*, TensorInfo>;
                 using TensorToInfoMap = std::unordered_map<descriptor::Tensor*, TensorInfo>;
@@ -170,27 +142,6 @@ namespace ngraph
                 TensorToInfoMap m_tensorToValueMap;
                 static const MLIRCompOpMap opDispatcher;
 
-                // REMOVE to backend
-
-                // Optimization level used by MLIR and LLVM compilers. It's based on LLVM CG
-                // optimization levels:
-                // enum Level {
-                //   None,        // -O0
-                //   Less,        // -O1
-                //   Default,     // -O2, -Os
-                //   Aggressive   // -O3
-                // };
-                // static llvm::CodeGenOpt::Level mlirOptLevel;
-
-                // LLVM target machine to be used by this MLIR compiler instance to retrieve
-                // information about target features.
-                // TODO: Note that, unfortunatelly, MLIR/OrcJIT execution engine creates its own
-                // target machine for compilation internally. This target machine is for non-JIT
-                // related stuff. We should change OrcJIT API so that we can pass an external target
-                // machine or configuration flags.
-                // TODO: Move target machine to external nGraph backend when multiple backends start
-                // to use MLIR.
-                //static std::unique_ptr<llvm::TargetMachine> targetMachine;
             };
         }
     }
