+++ src/contrib/mlir/tools/ngraph-opt/ngraph_opt.cpp
@@ -24,7 +24,7 @@
 /// small sequence of passes without running the whole compiler pipeline. Please, refer to
 /// ngraph_repo_path/tests/mlir/ for examples.
 
-#include "contrib/mlir/core/utils.hpp"
+#include "contrib/mlir/utils.hpp"
 #include "ngraph/check.hpp"
 
 #include <llvm/Support/CommandLine.h>
