+++ src/contrib/mlir/utils.hpp
@@ -16,14 +16,22 @@
 
 // NOTE: This file follows nGraph format style and MLIR naming convention since it does
 // not expose public API to the rest of nGraph codebase and heavily depends on MLIR API.
+#pragma once
 
-#include "utils.hpp"
+#include <mlir/IR/Module.h>
 
-#include "dialect/dialect.hpp"
-
-#include <mlir/IR/Dialect.h>
-
-void ngraph::runtime::ngmlir::initializeNGraphMLIR()
+namespace ngraph
 {
-    mlir::registerDialect<mlir::NGraphOpsDialect>();
-}
+    namespace runtime
+    {
+        namespace ngmlir
+        {
+            /// Common nGraph dialect initialization code. Used by nGraph compiler and tools that
+            /// require nGraph dialect initialization.
+            void initializeNGraphMLIR();
+
+            /// Helper to dump MLIR module into llvm::dbgs prepended by the message \p msg.
+            void dumpMlirModule(const std::string msg, mlir::ModuleOp module);
+        } // namespace ngmlir
+    }     // namespace runtime
+} // namespace ngraph
